Updated Backend Dir and Frontend for better scalability
