+++ backend/.env
@@ -1,0 +1,3 @@
+MONGO_URI=mongodb://localhost:27017/
+DATABASE_NAME=BusFinder
+PORT=5000
