+++ backend/models/Admin.js
@@ -1,0 +1,9 @@
+const mongoose = require("mongoose");
+
+const adminSchema = new mongoose.Schema({
+    username: { type: String, required: true, unique: true },
+    password: { type: String, required: true },
+    role: { type: String, required: true } // e.g., 'admin', 'kt-supervisor', 'pt-supervisor'
+});
+
+module.exports = mongoose.model("Admin", adminSchema);
