+++ backend/package.json
@@ -1,11 +1,13 @@
 {
   "dependencies": {
+    "bcrypt": "^5.1.1",
     "child_process": "^1.0.2",
     "cors": "^2.8.5",
     "dotenv": "^16.4.7",
     "express": "^4.21.2",
     "fs": "^0.0.1-security",
     "mongodb": "^6.13.1",
+    "mongoose": "^8.11.0",
     "multer": "^1.4.5-lts.1",
     "path": "^0.12.7"
   },
